+++ src/main/java/com/yuuko/core/Configuration.java
@@ -40,7 +40,7 @@ import java.util.*;
 public class Configuration {
     private static final Logger log = LoggerFactory.getLogger(Configuration.class);
 
-    public static final String VERSION = "2019-06-26";
+    public static final String VERSION = "2019-06-29";
     public static String AUTHOR;
     public static String AUTHOR_WEBSITE;
     public static String SUPPORT_GUILD;
@@ -58,6 +58,8 @@ public class Configuration {
     public static DiscordBotListAPI BOT_LIST;
     public static DivineAPI DIVINE_API;
     public static List<String> STANDARD_STRINGS;
+
+    public static boolean LOG_METRICS = true;
 
     /**
      * Loads all of the bots configurations. (Order DOES matter)