+++ src/main/java/com/yuuko/core/commands/developer/DeveloperModule.java
@@ -15,7 +15,8 @@ public class DeveloperModule extends Module {
             new SyncGuildsCommand(),
             new ReloadApiCommand(),
             new LavalinkNodeCommand(),
-            new ReloadDatabaseCommand()
+            new ReloadDatabaseCommand(),
+            new LogMetricsCommand()
     );
 
     public DeveloperModule(MessageEvent e) {