+++ src/main/java/com/yuuko/core/events/controllers/GenericGuildController.java
@@ -69,7 +69,7 @@ public class GenericGuildController {
         GuildFunctions.addGuild(e.getGuild());
 
         try {
-            if(e.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_READ, Permission.MESSAGE_WRITE)) {
+            if(e.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_READ, Permission.MESSAGE_WRITE, Permission.MESSAGE_EMBED_LINKS)) {
                 e.getGuild().getTextChannels().stream().filter(textChannel -> textChannel.getName().toLowerCase().contains("general")).findFirst().ifPresent(textChannel -> {
                     EmbedBuilder about = new EmbedBuilder()
                             .setAuthor(Configuration.BOT.getName() + "#" + Configuration.BOT.getDiscriminator(), null, Configuration.BOT.getAvatarUrl())