+++ src/main/java/com/yuuko/core/scheduler/jobs/TenSecondlyJob.java
@@ -1,5 +1,6 @@
 package com.yuuko.core.scheduler.jobs;
 
+import com.yuuko.core.Configuration;
 import com.yuuko.core.scheduler.Job;
 import com.yuuko.core.scheduler.tasks.UpdateMetricsTask;
 
@@ -14,6 +15,8 @@ public class TenSecondlyJob extends Job {
 
     @Override
     public void run() {
-        handleTask(updateMetricsTask);
+        if(Configuration.LOG_METRICS) {
+            handleTask(updateMetricsTask);
+        }
     }
 }